requirements.txt
asgiref==3.6.0
Django==4.1.7
django-jazzmin==2.6.0
django-quill-editor==0.1.40
Pillow==9.5.0
sqlparse==0.4.3
tzdata==2023.3
